Responses to Matt's comments 2016/6/2
Date of drawing ?

Isn't it 6 RF/SNAP ?

What about the Post Amplifier Chassis ?

PDU
  includes network controlled power strip and
  SNAP PSU(s), what else ??

Is VDC input power to each SNAP independently
controlled?
if so, how ?
perhaps, no need; if rare case when actual
power cycling required then all 4 SNAP will be
power cycled ?

Not shown are the sensor inputs, not on GbE,
to the monitor computer - right.
ah - ha - This isn't shown because this is the (digital) block diagram - so never mind.

Is 8 point switch enough ports ?
  4 snap + WRE + CPU + PDU + 1 spare
  where the spare can be for the tech support laptop
  etc.

Is the SMF 4:1 mux/demux purely passive ?
or does it include amplifiers ?

Add a wire from PDU heading off-page to the
blower ?  Again may be not shown because of
(digital) block diagram.

the 2 grey blocks vs 1 black box for the
SNAP network IO is confusing.